Implement health system: Add Health class, update character and enemy scripts to use integer health values, and modify projectile damage handling to support integer types.

+++ Assets/Passive Improvments/Scripts/PassiveImprovment.cs
@@ -16,8 +16,8 @@ namespace PassiveImprovments
 
         public int duration;
         public float radius;
-        public float hpGain;
-        public float hpRegen;
+        public int hpGain;
+        public int hpRegen;
         public float cooldown;
 
         public float activatingTime=0.3f;
